requirements.txt
aiogram>=3.0.0
aiohttp>=3.8.0
aiosqlite>=0.18.0
python-dotenv>=1.0.0
beautifulsoup4>=4.11.0
requests>=2.28.0
selenium>=4.15.0
webdriver-manager>=3.8.6
